+++ include/System/Animation.hpp
@@ -7,36 +7,36 @@ class Animation
 {
 public:
     Animation();
-    Animation(size_t sheetW, size_t sheetH, size_t spriteW, size_t spriteH, size_t frameCount, float speed);
-    void setTextureID(size_t textureID);
+    Animation(uint32_t sheetW, uint32_t sheetH, uint32_t spriteW, uint32_t spriteH, uint32_t frameCount, float speed);
+    void setTextureID(uint32_t textureID);
     void setDone() {_done=true;}
     void setCenter(sf::Vector2f center) {_spriteCenter=center;}
     void resetTime(float time);
 
     sf::IntRect  getTexCoord(float time);
-    sf::Vector2f getSpriteSize()   const {return sf::Vector2f(_spriteW, _spriteH);}
+    sf::Vector2f getSpriteSize()   const {return sf::Vector2f(static_cast<float>(_spriteW), static_cast<float>(_spriteH));}
     sf::Vector2f getSpriteCenter() const {return _spriteCenter;}
-    size_t       getTexture()      const {return _textureID;}
-    size_t       getCurrentIndex(float time) const {return _animationSpeed*(time-_startTime);}
+	uint32_t       getTexture()      const {return _textureID;}
+	uint32_t       getCurrentIndex(float time) const {return static_cast<size_t>(_animationSpeed*(time-_startTime));}
 
     void applyOnQuad(sf::VertexArray& quad, float time);
 
     bool isDone() const;
 
 private:
-    size_t _sheetW;
-    size_t _sheetH;
-    size_t _spriteW;
-    size_t _spriteH;
+    uint32_t _sheetW;
+    uint32_t _sheetH;
+    uint32_t _spriteW;
+    uint32_t _spriteH;
 
     float  _startTime;
 
-    size_t _frameCount;
+	uint32_t _frameCount;
     float  _animationSpeed;
     bool   _done;
 
     sf::Vector2f _spriteCenter;
-    size_t       _textureID;
+	uint32_t       _textureID;
 
 };
